First commit with two working examples:
 * test1_image:
Load image from file with OpenCV, it converts it to Raylib format and the to raylib texture
Then a new Raylib windo is created and the texture is shown at 60fps

 * test2_camera:
Load video from file with OpenCV, and using Opecv functions it detects moving objects (very basic algo).
The result is stored as Raylib rectangles and displayed together with the frame in a raylib window.